fix an ui issue

+++ app/src/main/java/com/example/check24/overview/data/RemoteDataSourceImpl.kt
@@ -32,5 +32,5 @@ class RemoteDataSourceImpl @Inject constructor(val api: ProductOverviewApi) : Re
 
 fun Product.toProductEntity(isFooter : Boolean) = ProductEntity(0,
     isFooter, available, color, colorCode,
-    description, imageURL, longDescription, name, price.currency, price.value.toString(), rating, DateUtil.getDateFromMiliSecond(releaseDate.toLong()), type, false
+    description, imageURL, longDescription, name, price.currency, price.value.toString(), rating.toFloat(), DateUtil.getDateFromMiliSecond(releaseDate.toLong()), type, false
 )
